Modified Cost Function
Added an entanglement layer. Modified the cost function to be far more computationally efficient by replacing the for loop with a matrix operation. This greatly lowers exectution times, and will allow me to run far more experiments then before. Also removed the last cell, was not relevant for this project.